--- server/crud.go
@@ -1,10 +1,0 @@
-package server
-
-import "github.com/gorethink/gorethink"
-
-// CreateTable create table
-func CreateTable(name string) {
-	gorethink.TableCreate("foo").
-		Wait().
-		RunWrite(GetSession())
-}
