+++ server/middleware/metrics.go
@@ -1,0 +1,11 @@
+package middleware
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+// MetricsMiddleware accumulate request metrics
+func MetricsMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+	}
+}
