+++ server/middleware/middleware.go
@@ -1,0 +1,12 @@
+package middleware
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+// Middleware base
+func Middleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.Header("Content-Type", "application/json; charset=utf-8")
+	}
+}
